# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/evaluations/count/courses/{courseId}/invisible": {
            "get": {
                "description": "根据课程ID计算该课程的不可见评价数量。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "不可见课评数",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/count/mine": {
            "get": {
                "description": "根据用户ID和课评状态分类统计用户的课评数量。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "用户课评数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "课评状态",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/list/all": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "课评列表[广场]",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前ID",
                        "name": "cur_evaluation_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "课评数量限制",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用于过滤课评的课程性质（可选）",
                        "name": "property",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/evaluation.EvaluationVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/list/courses/{courseId}": {
            "get": {
                "description": "根据课程ID获取课程评价列表，支持分页。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "课评列表[指定课程]",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前课评ID",
                        "name": "cur_evaluation_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "返回课评的最大数量，上限为100",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/evaluation.EvaluationVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/list/mine": {
            "get": {
                "description": "根据课程ID获取课程评价列表，支持分页。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "课评列表[我的历史]",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前评估ID，用于分页",
                        "name": "cur_evaluation_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "上限为100",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "课评状态: Public/Private/Folded",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/evaluation.EvaluationVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/save": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "发布课评",
                "parameters": [
                    {
                        "description": "发布课评请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/evaluation.SaveReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/{evaluationId}/detail": {
            "get": {
                "description": "根据课评ID获取详情，详情包括标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "课评详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课评ID",
                        "name": "evaluationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/evaluation.EvaluationVo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/{evaluationId}/endorse": {
            "post": {
                "description": "根据评价ID支持或反对指定的评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "支持或反对课评",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评价ID",
                        "name": "evaluationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "支持或反对的请求体，stance为态度标识，-1反对，1支持，0表示无可用于取消表态",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/evaluation.EndorseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/ginx.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/evaluations/{evaluationId}/status": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课评"
                ],
                "summary": "变更课评状态",
                "parameters": [
                    {
                        "description": "变更课评状态请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/evaluation.UpdateStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ginx.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "evaluation.EndorseReq": {
            "type": "object",
            "properties": {
                "stance": {
                    "type": "integer"
                }
            }
        },
        "evaluation.EvaluationVo": {
            "type": "object",
            "properties": {
                "assessments": {
                    "description": "考核方式，支持多选",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "content": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "ctime": {
                    "type": "integer"
                },
                "features": {
                    "description": "课程特点，支持多选",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "publisher_id": {
                    "type": "integer"
                },
                "stance": {
                    "type": "integer"
                },
                "star_rating": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_comment_count": {
                    "type": "integer"
                },
                "total_oppose_count": {
                    "type": "integer"
                },
                "total_support_count": {
                    "type": "integer"
                },
                "utime": {
                    "type": "integer"
                }
            }
        },
        "evaluation.SaveReq": {
            "type": "object",
            "properties": {
                "assessments": {
                    "description": "考核方式，支持多选",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "content": {
                    "description": "评价的内容",
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "features": {
                    "description": "课程特点，支持多选",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "star_rating": {
                    "description": "1，2，3，4，5",
                    "type": "integer"
                },
                "status": {
                    "description": "可见性：Public/Private",
                    "type": "string"
                }
            }
        },
        "evaluation.UpdateStatusReq": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "ginx.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码，非 0 表示失败",
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "description": "错误或成功 描述",
                    "type": "string"
                }
            }
        },
...
    }
}